djari_clothes/requirements.txt
asgiref==3.5.2
cycler==0.11.0
Django==3.2.4
django-cors-headers==3.10.0
django-model-utils==4.0.0
djangorestframework==3.11.0
fonttools==4.33.3
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
